Move schema loading logic
