+++ packages/graphql-codegen-cli/src/load.ts
@@ -1,3 +1,5 @@
+import { Types } from 'graphql-codegen-core';
+import { GraphQLSchema } from 'graphql';
 import { loadDocumentsSources } from './loaders/documents/document-loader';
 import { documentsFromGlobs } from './utils/documents-glob';
 import { IntrospectionFromFileLoader } from './loaders/schema/introspection-from-file';
@@ -5,7 +7,6 @@ import { IntrospectionFromUrlLoader } from './loaders/schema/introspection-from-
 import { SchemaFromTypedefs } from './loaders/schema/schema-from-typedefs';
 import { SchemaFromExport } from './loaders/schema/schema-from-export';
 import { DetailedError } from './errors';
-import { CLIOptions, createConfigFromOldCli } from './old-cli-config';
 
 const schemaHandlers = [
   new IntrospectionFromUrlLoader(),
@@ -14,17 +15,27 @@ const schemaHandlers = [
   new SchemaFromExport()
 ];
 
-export async function loadSchema(pointToSchema: string, options: CLIOptions & { [key: string]: any }) {
+export async function loadSchema(schemaDef: Types.Schema, config: Types.Config): Promise<GraphQLSchema> {
   for (const handler of schemaHandlers) {
+    let pointToSchema: string = null;
+    let options: any = {};
+
+    if (typeof schemaDef === 'string') {
+      pointToSchema = schemaDef as string;
+    } else if (typeof schemaDef === 'object') {
+      pointToSchema = Object.keys(schemaDef)[0];
+      options = schemaDef[pointToSchema];
+    }
+
     if (await handler.canHandle(pointToSchema)) {
-      return handler.handle(pointToSchema, createConfigFromOldCli(options), null);
+      return handler.handle(pointToSchema, config, options);
     }
   }
 
   throw new DetailedError(
     'Failed to load schema',
     `
-    Failed to load schema from ${pointToSchema}.
+    Failed to load schema: ${schemaDef}.
 
     GraphQL Code Generator supports:
       - esmodules and commonjs exports
